requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.7
distlib==0.3.9
distro==1.9.0
eval_type_backport==0.2.2
fal==1.12.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
google==3.0.0
google-auth==2.38.0
google-genai==1.10.0
grpc-interceptor==0.15.4
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
isolate==0.16.1
isolate_proto==0.7.2
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.2
numpy==2.2.4
openai==1.72.0
opentelemetry-api==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
portalocker==2.10.1
prometheus_client==0.21.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pywin32==310
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-argparse==1.7.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
starlette_exporter==0.23.0
structlog==22.3.0
tabulate==0.9.0
tblib==3.1.0
text-unidecode==1.3
together==1.5.5
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
